player2many: add error control to Kurento API calls
Some of the JS API calls are ready for async/await style code, but some
other are still using the old callback-based style, so they need to be
Promisified in order to use them with async calls. Otherwise, in case of
server error, nothing with be caught in the JS side.